refactor: clean projet ouds_core
