+++ app/lib/ui/components/checkbox/checkbox_item_demo_screen.dart
@@ -12,9 +12,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/checkbox/ouds_checkbox_item.dart';
-import 'package:ouds_core/components/lists/ouds_list_switch.dart';
 import 'package:ouds_core/components/ouds_colored_box.dart';
-import 'package:ouds_core/components/sheets_bottom/ouds_sheets_bottom.dart';
 import 'package:ouds_flutter_demo/l10n/app_localizations.dart';
 import 'package:ouds_flutter_demo/main_app_bar.dart';
 import 'package:ouds_flutter_demo/ui/components/control_item/control_item_code_generator.dart';
@@ -24,8 +22,10 @@ import 'package:ouds_flutter_demo/ui/components/control_item/control_item_enum.d
 import 'package:ouds_flutter_demo/ui/theme/theme_controller.dart';
 import 'package:ouds_flutter_demo/ui/utilities/code.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_section.dart';
+import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_switch.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_textfield.dart';
 import 'package:ouds_flutter_demo/ui/utilities/detail_screen_header.dart';
+import 'package:ouds_flutter_demo/ui/utilities/sheets_bottom/ouds_sheets_bottom.dart';
 import 'package:provider/provider.dart';
 
 /// This screen displays a checkbox demo and allows customization of checkbox properties.
@@ -184,7 +184,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
 
     return CustomizableSection(
       children: [
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_components_controlItem_icon_label,
           value: customizationState!.hasIcon,
           onChanged: (value) {
@@ -193,7 +193,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
             });
           },
         ),
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_components_controlItem_divider_label,
           value: customizationState.hasDivider,
           onChanged: (value) {
@@ -202,7 +202,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
             });
           },
         ),
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_components_controlItem_reversed_label,
           value: customizationState.hasReversed,
           onChanged: (value) {
@@ -211,7 +211,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
             });
           },
         ),
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_common_enabled_label,
           value: customizationState.hasEnabled,
           onChanged: customizationState.isEnabledWhenError || customizationState.isEnabledWhenReadOnly
@@ -222,7 +222,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                   });
                 },
         ),
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_components_controlItem_readOnly_label,
           value: customizationState.hasReadOnly,
           onChanged: customizationState.isReadOnlyWhenError || customizationState.isReadOnlyWhenEnabled
@@ -233,7 +233,7 @@ class _CustomizationContentState extends State<_CustomizationContent> {
                   });
                 },
         ),
-        OudsListSwitch(
+        CustomizableSwitch(
           title: context.l10n.app_components_common_error_label,
           value: customizationState.hasError,
           onChanged: customizationState.isErrorWhenEnabled || customizationState.isErrorWhenReadOnly
