+++ app/lib/ui/tokens/token_detail_screen.dart
@@ -12,7 +12,6 @@
 
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:ouds_core/components/lists/ouds_list_item.dart';
 import 'package:ouds_flutter_demo/main_app_bar.dart';
 import 'package:ouds_flutter_demo/ui/tokens/token_entities.dart';
 import 'package:ouds_flutter_demo/ui/utilities/display_image.dart';
@@ -81,11 +80,10 @@ class VariantEntry extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return OudsListItem(
-      title: variant.title,
-      subtitle: variant.technicalName,
-      image: const Icon(Icons.play_circle_outline),
-      onPressed: () {
+    return ListTile(
+      title: Text(variant.title),
+      subtitle: Text(variant.technicalName),
+      onTap: () {
         Get.to(variant.screen);
       },
     );
